Separate reusable installer and update UI
