--- Source/components/UpdatePromptComponent.h
@@ -1,574 +1,0 @@
-#pragma once
-
-#include <JuceHeader.h>
-#include <melatonin_blur/melatonin_blur.h>
-#include "../CommonPluginProcessor.h"
-#include "InstallFlowHelpers.h"
-#include "../LookAndFeel.h"
-
-class UpdatePromptComponent : public juce::Component {
-public:
-    explicit UpdatePromptComponent(CommonAudioProcessor& processorToUse)
-        : processor(processorToUse), closeButton("Dismiss update", osci::Colours::text(), osci::Colours::text()) {
-        setVisible (false);
-        setAlwaysOnTop (true);
-
-        titleLabel.setFont (juce::Font (juce::FontOptions (17.0f, juce::Font::bold)));
-        titleLabel.setColour (juce::Label::textColourId, juce::Colours::white);
-        titleLabel.setJustificationType (juce::Justification::centredLeft);
-
-        detailLabel.setFont (juce::Font (juce::FontOptions (14.5f, juce::Font::plain)));
-        detailLabel.setColour (juce::Label::textColourId, juce::Colours::white.withAlpha (0.78f));
-        detailLabel.setJustificationType (juce::Justification::centredLeft);
-        detailLabel.setMinimumHorizontalScale (0.75f);
-
-        primaryButton.setColour (juce::TextButton::buttonColourId, osci::Colours::accentColor());
-        primaryButton.setColour (juce::TextButton::textColourOffId, osci::Colours::veryDark());
-        secondaryButton.setColour (juce::TextButton::buttonColourId, osci::Colours::veryDark().brighter (0.12f));
-
-        primaryButton.onClick = [this] {
-            if (mode == Mode::Ready) {
-                installUpdate();
-            } else if (mode == Mode::PendingInstallFailed) {
-                retryPendingInstall();
-            } else if (mode == Mode::Available || mode == Mode::Failed) {
-                downloadUpdate();
-            }
-        };
-
-        secondaryButton.onClick = [this] { dismissFor48Hours(); };
-        closeButton.setPaintsBackground(true);
-        closeButton.setIconPadding(8);
-        closeButton.onClick = [this] { dismissFor48Hours(); };
-
-        addAndMakeVisible (titleLabel);
-        addAndMakeVisible (detailLabel);
-        addAndMakeVisible(closeButton);
-        addAndMakeVisible (primaryButton);
-        addAndMakeVisible (secondaryButton);
-        addChildComponent (progressBar);
-        addChildComponent(notification);
-        notification.onDismiss = [this] { hidePrompt(); };
-    }
-
-    std::function<void()> onLicenseRequired;
-
-    void scheduleInitialCheck() {
-        auto safeThis = juce::Component::SafePointer<UpdatePromptComponent> (this);
-        juce::Timer::callAfterDelay (1800, [safeThis] {
-            if (safeThis != nullptr) {
-                safeThis->checkForUpdate (false);
-            }
-        });
-    }
-
-    bool showPendingInstallStatusIfNeeded() {
-        osci::PendingInstall pending (processor.getProductSlug());
-        auto marker = pending.load();
-        if (! marker.has_value()) {
-            return false;
-        }
-
-        if (osci::PendingInstall::isResolvedByRunningVersion (*marker, JucePlugin_VersionString)) {
-            pending.clear();
-            showInstallSucceeded (*marker);
-            return true;
-        }
-
-        pending.clear();
-        showPendingInstallFailed (*marker);
-        return true;
-    }
-
-    int getPreferredWidth() const noexcept {
-        return mode == Mode::InstallSucceeded ? 410 : 560;
-    }
-
-    int getPreferredHeight() const noexcept {
-        if (mode == Mode::Downloading) {
-            return 234;
-        }
-
-        return mode == Mode::InstallSucceeded ? 88 : 184;
-    }
-
-    void paint (juce::Graphics& g) override {
-        if (mode == Mode::InstallSucceeded) {
-            return;
-        }
-
-        const auto bounds = getCardBounds().toFloat();
-
-        juce::Path shadowPath;
-        shadowPath.addRoundedRectangle (bounds, cornerRadius);
-        shadow.render (g, shadowPath);
-
-        g.setColour (osci::Colours::veryDark().withAlpha (0.985f));
-        g.fillRoundedRectangle (bounds, cornerRadius);
-        g.setColour (osci::Colours::accentColor().withAlpha (0.72f));
-        g.drawRoundedRectangle (bounds.reduced (0.5f), cornerRadius, 1.2f);
-    }
-
-    void resized() override {
-        if (mode == Mode::InstallSucceeded) {
-            titleLabel.setBounds({});
-            detailLabel.setBounds({});
-            closeButton.setBounds({});
-            primaryButton.setBounds({});
-            secondaryButton.setBounds({});
-            progressBar.setBounds({});
-            notification.setBounds(getLocalBounds());
-            return;
-        }
-
-        notification.setBounds({});
-        const auto cardBounds = getCardBounds();
-        closeButton.setBounds(cardBounds.getRight() - 44, cardBounds.getY() + 14, 28, 28);
-
-        auto area = cardBounds.reduced (28, 22);
-        auto titleArea = area.removeFromTop(24);
-        titleArea.removeFromRight(38);
-        titleLabel.setBounds(titleArea);
-        area.removeFromTop (8);
-        detailLabel.setBounds (area.removeFromTop (24));
-        area.removeFromTop (12);
-
-        if (progressBar.isVisible()) {
-            progressBar.setBounds (area.removeFromTop (34));
-            area.removeFromTop (16);
-        }
-
-        auto buttonRow = area.removeFromBottom (40);
-        if (secondaryButton.isVisible()) {
-            secondaryButton.setBounds (buttonRow.removeFromRight (112));
-            buttonRow.removeFromRight (12);
-        } else {
-            secondaryButton.setBounds ({});
-        }
-        primaryButton.setBounds (buttonRow.removeFromRight (128));
-    }
-
-private:
-    class UpdateProgressBar final : public juce::Component {
-    public:
-        void setProgress (double newProgress) {
-            progress = juce::jlimit (0.0, 1.0, newProgress);
-            repaint();
-        }
-
-        void paint (juce::Graphics& g) override {
-            const auto bounds = getLocalBounds().toFloat().reduced (0.5f);
-            const auto fillWidth = bounds.getWidth() * static_cast<float>(progress);
-
-            g.setColour (osci::Colours::veryDark().brighter (0.12f));
-            g.fillRoundedRectangle (bounds, 8.0f);
-
-            if (fillWidth > 1.0f) {
-                g.setColour (osci::Colours::accentColor().withAlpha (0.92f));
-                g.fillRoundedRectangle (bounds.withWidth (fillWidth), 8.0f);
-            }
-
-            g.setColour (juce::Colours::white.withAlpha (0.92f));
-            g.setFont (juce::FontOptions (13.0f, juce::Font::bold));
-            g.drawText (juce::String (juce::roundToInt (progress * 100.0)) + "%",
-                        getLocalBounds(),
-                        juce::Justification::centred);
-        }
-
-    private:
-        double progress = 0.0;
-    };
-
-    enum class Mode {
-        Hidden,
-        Available,
-        Downloading,
-        Ready,
-        Failed,
-        InstallSucceeded,
-        PendingInstallFailed,
-    };
-
-    CommonAudioProcessor& processor;
-    juce::Label titleLabel;
-    juce::Label detailLabel;
-    osci::CloseButton closeButton;
-    juce::TextButton primaryButton { "Update" };
-    juce::TextButton secondaryButton { "Later" };
-    UpdateProgressBar progressBar;
-    osci::NotificationComponent notification;
-    std::optional<osci::VersionInfo> availableVersion;
-    std::optional<osci::PendingInstallMarker> pendingInstallRetryMarker;
-    juce::File downloadedFile;
-#if JUCE_LINUX
-    std::unique_ptr<osci::LinuxUpdateInstallThread> linuxInstallThread;
-#endif
-    Mode mode = Mode::Hidden;
-    bool busy = false;
-    melatonin::DropShadow shadow {
-        { juce::Colours::black.withAlpha (0.62f), 20, { 0, 8 }, 0 },
-        { juce::Colours::black.withAlpha (0.42f), 8, { 0, 2 }, 0 }
-    };
-
-    static constexpr int shadowPadding = 16;
-    static constexpr float cornerRadius = 10.0f;
-
-    juce::Rectangle<int> getCardBounds() const {
-        return getLocalBounds().reduced (shadowPadding);
-    }
-
-    void checkForUpdate (bool showErrors) {
-        if (busy || mode == Mode::InstallSucceeded || mode == Mode::PendingInstallFailed) {
-            return;
-        }
-
-        busy = true;
-        const auto product = processor.getProductSlug();
-        const auto currentVersion = juce::String (JucePlugin_VersionString);
-        const auto track = selectedReleaseTrack();
-        const auto variant = desiredUpdateVariant();
-        auto foundVersion = std::make_shared<std::optional<osci::VersionInfo>>();
-        auto result = std::make_shared<juce::Result> (juce::Result::ok());
-        auto safeThis = juce::Component::SafePointer<UpdatePromptComponent> (this);
-
-        juce::Thread::launch ([safeThis, foundVersion, result, product, currentVersion, track, variant, showErrors] {
-            osci::UpdateChecker checker;
-            *foundVersion = checker.checkForUpdate (product, currentVersion, track, variant);
-            *result = checker.getLastResult();
-            if (result->failed()) {
-                *result = failWithContext ("Could not check for updates", *result);
-            }
-
-            juce::MessageManager::callAsync ([safeThis, foundVersion, result, showErrors] {
-                if (safeThis == nullptr) {
-                    return;
-                }
-
-                safeThis->busy = false;
-                if (result->failed()) {
-                    if (showErrors) {
-                        safeThis->showFailure (result->getErrorMessage());
-                    }
-                    return;
-                }
-
-                if (! foundVersion->has_value()) {
-                    return;
-                }
-
-                if (safeThis->isDismissed ((*foundVersion)->semver)) {
-                    return;
-                }
-
-                safeThis->showAvailable (**foundVersion);
-            });
-        });
-    }
-
-    void showAvailable (const osci::VersionInfo& version) {
-        availableVersion = version;
-        downloadedFile = juce::File();
-        mode = Mode::Available;
-        progressBar.setVisible (false);
-        titleLabel.setText ("Update available", juce::dontSendNotification);
-        detailLabel.setText (updateAvailableDetailText (version), juce::dontSendNotification);
-        primaryButton.setButtonText ("Update");
-        secondaryButton.setButtonText ("Later");
-        secondaryButton.setVisible (true);
-        primaryButton.setEnabled (true);
-        secondaryButton.setEnabled (true);
-        setVisible (true);
-        refreshParentLayout();
-        repaint();
-    }
-
-    void showFailure (juce::StringRef message) {
-        mode = Mode::Failed;
-        progressBar.setVisible (false);
-        titleLabel.setText ("Update failed", juce::dontSendNotification);
-        detailLabel.setText (message, juce::dontSendNotification);
-        primaryButton.setButtonText ("Retry");
-        secondaryButton.setButtonText ("Later");
-        secondaryButton.setVisible (true);
-        primaryButton.setEnabled (true);
-        secondaryButton.setEnabled (true);
-        setVisible (true);
-        refreshParentLayout();
-        repaint();
-    }
-
-    void downloadUpdate() {
-        if (! availableVersion.has_value() || busy) {
-            return;
-        }
-
-        const auto token = processor.licenseManager.getCachedToken();
-        if (isPremiumDownloadWithoutToken (*availableVersion, token)) {
-            if (onLicenseRequired) {
-                onLicenseRequired();
-            }
-            return;
-        }
-
-        busy = true;
-        mode = Mode::Downloading;
-        progressBar.setProgress (0.0);
-        progressBar.setVisible (true);
-        titleLabel.setText ("Downloading update", juce::dontSendNotification);
-        detailLabel.setText ("Preparing a verified installer...", juce::dontSendNotification);
-        primaryButton.setEnabled (false);
-        secondaryButton.setEnabled (false);
-        refreshParentLayout();
-
-        auto downloaded = std::make_shared<juce::File>();
-        auto result = std::make_shared<juce::Result> (juce::Result::ok());
-        const auto version = *availableVersion;
-        const auto product = processor.getProductSlug();
-        auto safeThis = juce::Component::SafePointer<UpdatePromptComponent> (this);
-
-        juce::Thread::launch ([safeThis, downloaded, result, version, token, product] {
-            osci::Downloader::Config config;
-            config.downloadDirectory = osci::HardwareInfo::getDefaultStorageDirectory (product).getChildFile ("downloads");
-            osci::Downloader downloader (config);
-            *result = downloader.downloadAndVerify (version, token, [safeThis] (double fraction, juce::int64) {
-                juce::MessageManager::callAsync ([safeThis, fraction] {
-                    if (safeThis != nullptr && fraction >= 0.0) {
-                        safeThis->progressBar.setProgress (fraction);
-                    }
-                });
-            });
-
-            if (result->wasOk()) {
-                *downloaded = downloader.getDownloadedFile();
-            } else {
-                *result = failWithContext ("Could not download update", *result);
-            }
-
-            juce::MessageManager::callAsync ([safeThis, downloaded, result] {
-                if (safeThis == nullptr) {
-                    return;
-                }
-
-                safeThis->busy = false;
-                if (result->failed()) {
-                    safeThis->showFailure (result->getErrorMessage());
-                    return;
-                }
-
-                safeThis->downloadedFile = *downloaded;
-                safeThis->mode = Mode::Ready;
-                safeThis->progressBar.setProgress (1.0);
-                safeThis->progressBar.setVisible (false);
-                safeThis->titleLabel.setText ("Update ready", juce::dontSendNotification);
-                safeThis->detailLabel.setText ("The verified installer has downloaded.", juce::dontSendNotification);
-                safeThis->primaryButton.setButtonText ("Install");
-                safeThis->secondaryButton.setButtonText ("Later");
-                safeThis->secondaryButton.setVisible (true);
-                safeThis->primaryButton.setEnabled (true);
-                safeThis->secondaryButton.setEnabled (true);
-                safeThis->refreshParentLayout();
-                safeThis->repaint();
-            });
-        });
-    }
-
-    void installUpdate() {
-        if (! downloadedFile.existsAsFile()) {
-            return;
-        }
-
-        const auto file = downloadedFile;
-        const auto version = availableVersion;
-        const auto product = processor.getProductSlug();
-        const auto currentVersion = juce::String (JucePlugin_VersionString);
-        auto safeThis = juce::Component::SafePointer<UpdatePromptComponent> (this);
-
-        osci::InstallPrompt::showConfirmation ({
-            this,
-            [safeThis, file, version, product, currentVersion] {
-#if JUCE_LINUX
-                if (!version.has_value() || safeThis == nullptr) {
-                    return;
-                }
-
-                safeThis->busy = true;
-                safeThis->mode = Mode::Downloading;
-                safeThis->progressBar.setProgress (0.0);
-                safeThis->progressBar.setVisible (true);
-                safeThis->titleLabel.setText ("Installing update", juce::dontSendNotification);
-                safeThis->primaryButton.setEnabled (false);
-                safeThis->secondaryButton.setEnabled (false);
-                safeThis->refreshParentLayout();
-
-                osci::LinuxUpdateInstallThread::Request request;
-                request.archive = file;
-                request.version = *version;
-                request.currentVersion = currentVersion;
-                request.manifest = osci::linuxInstallManifest (product);
-                request.iconPng = osci::linuxProductIconPng();
-                request.progress = [safeThis] (double fraction, juce::StringRef stage) {
-                    juce::MessageManager::callAsync ([safeThis, fraction, stage = juce::String (stage)] {
-                        if (safeThis != nullptr) {
-                            safeThis->progressBar.setProgress (fraction);
-                            safeThis->detailLabel.setText (stage, juce::dontSendNotification);
-                        }
-                    });
-                };
-                request.completion = [safeThis, version] (juce::Result result, osci::LinuxInstaller::Report report) {
-                    if (safeThis == nullptr) {
-                        return;
-                    }
-
-                    safeThis->busy = false;
-                    if (result.failed()) {
-                        safeThis->showFailure (result.getErrorMessage());
-                        return;
-                    }
-
-                    auto detail = "Version " + version->semver + " is installed. Restart the app or plugin host to use it.";
-                    if (!report.warnings.isEmpty()) {
-                        detail << " " << report.warnings.joinIntoString (" ");
-                    }
-                    safeThis->mode = Mode::InstallSucceeded;
-                    safeThis->progressBar.setVisible (false);
-                    safeThis->notification.setNotification ("Update installed", detail);
-                    safeThis->refreshParentLayout();
-                    safeThis->repaint();
-                };
-                safeThis->linuxInstallThread = osci::UpdateInstaller::installLinuxAsync (std::move (request));
-                if (safeThis->linuxInstallThread == nullptr) {
-                    safeThis->busy = false;
-                    safeThis->showFailure ("Could not start the Linux installer thread");
-                }
-#else
-                const auto result = osci::UpdateInstaller::launchWithPendingMarker ({ file, version, product, currentVersion });
-                if (result.failed()) {
-                    osci::InstallPrompt::showError (safeThis.getComponent(), "Install Update", result.getErrorMessage());
-                    return;
-                }
-
-                if (safeThis != nullptr) {
-                    safeThis->hidePrompt();
-                }
-#endif
-            }
-        });
-    }
-
-    void dismissFor48Hours() {
-        if (availableVersion.has_value()) {
-            osci::UpdateSettings(processor.getProductSlug()).dismiss (availableVersion->semver, currentTimeSeconds());
-        }
-
-        hidePrompt();
-    }
-
-    bool isDismissed (const juce::String& semver) const {
-        return osci::UpdateSettings(processor.getProductSlug()).isDismissed (semver, currentTimeSeconds());
-    }
-
-    osci::ReleaseTrack selectedReleaseTrack() const {
-        return osci::UpdateSettings(processor.getProductSlug()).releaseTrack();
-    }
-
-    juce::String desiredUpdateVariant() const {
-        if (processor.licenseManager.hasPremium()) {
-            return "premium";
-        }
-
-        return compiledVariant();
-    }
-
-    juce::String updateAvailableDetailText (const osci::VersionInfo& version) const {
-        if (isPremiumDownloadWithoutToken (version, processor.licenseManager.getCachedToken())) {
-            return "Version " + version.semver + " is available. A license key is required to download this premium update.";
-        }
-
-        return "Version " + version.semver + " is ready to install.";
-    }
-
-    static bool isPremiumDownloadWithoutToken (const osci::VersionInfo& version, juce::StringRef token) {
-        return version.variant == "premium" && token.isEmpty();
-    }
-
-    static juce::String compiledVariant() {
-#if OSCI_PREMIUM
-        return "premium";
-#else
-        return "free";
-#endif
-    }
-
-    static double currentTimeSeconds() {
-        return static_cast<double> (juce::Time::currentTimeMillis()) / 1000.0;
-    }
-
-    static juce::Result failWithContext (juce::StringRef context, const juce::Result& result) {
-        return juce::Result::fail (juce::String (context) + ": " + result.getErrorMessage());
-    }
-
-    void showInstallSucceeded (const osci::PendingInstallMarker& marker) {
-        availableVersion.reset();
-        pendingInstallRetryMarker.reset();
-        downloadedFile = juce::File();
-        mode = Mode::InstallSucceeded;
-        progressBar.setVisible (false);
-        notification.setNotification("Update installed", "Version " + marker.targetVersion + " is now running.");
-        setVisible (true);
-        refreshParentLayout();
-        repaint();
-    }
-
-    void showPendingInstallFailed (const osci::PendingInstallMarker& marker) {
-        pendingInstallRetryMarker = marker;
-        availableVersion = marker.toVersionInfo();
-        downloadedFile = marker.artifactPath.isNotEmpty() ? juce::File (marker.artifactPath) : juce::File();
-        mode = Mode::PendingInstallFailed;
-        progressBar.setVisible (false);
-        titleLabel.setText ("Install did not complete", juce::dontSendNotification);
-        detailLabel.setText ("Version " + marker.targetVersion + " was not installed.", juce::dontSendNotification);
-        primaryButton.setButtonText ("Retry");
-        secondaryButton.setButtonText ("Later");
-        secondaryButton.setVisible (true);
-        primaryButton.setEnabled (true);
-        secondaryButton.setEnabled (true);
-        setVisible (true);
-        refreshParentLayout();
-        repaint();
-    }
-
-    void retryPendingInstall() {
-        if (! pendingInstallRetryMarker.has_value()) {
-            return;
-        }
-
-        availableVersion = pendingInstallRetryMarker->toVersionInfo();
-
-        if (osci::PendingInstall::validateArtifact (*pendingInstallRetryMarker).wasOk()) {
-            downloadedFile = juce::File (pendingInstallRetryMarker->artifactPath);
-            installUpdate();
-            return;
-        }
-
-        downloadedFile = juce::File();
-        downloadUpdate();
-    }
-
-    void hidePrompt() {
-        mode = Mode::Hidden;
-        notification.setVisible(false);
-        setVisible (false);
-        refreshParentLayout();
-    }
-
-    void refreshParentLayout() {
-        auto* parent = getParentComponent();
-        if (parent != nullptr) {
-            parent->resized();
-        } else {
-            resized();
-        }
-    }
-
-    JUCE_DECLARE_NON_COPYABLE_WITH_LEAK_DETECTOR (UpdatePromptComponent)
-};
